refactor: global single name templating package
